backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.14.3
pytest==8.1.1
